@@ -8,28 +8,28 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
-  html: ^0.14.0+4
-  http: ^0.12.2
-  image_picker: ^0.6.7+15
-  intl: ^0.16.1
-  lottie: ^0.7.0+1
-  path: ^1.7.0
-  provider: ^4.3.2+3
-  quiver: ^2.1.5
-  settings_ui: ^0.5.0
-  shared_preferences: ^0.5.12+4
-  sqflite: ^1.3.2+1
+  cupertino_icons: ^1.0.3
+  html: ^0.15.0
+  http: ^0.13.3
+  image_picker: ^0.7.5+3
+  intl: ^0.17.0
+  lottie: ^1.0.1
+  path: ^1.8.0
+  provider: ^5.0.0
+  quiver: ^3.0.1
+  settings_ui: ^1.0.0
+  shared_preferences: ^2.0.6
+  sqflite: ^2.0.0+3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.8.1
+  flutter_launcher_icons: ^0.9.0
 
 flutter_icons:
   android: true
