@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 ///[QuestionPlaceholder] - displays a question.
 class QuestionPlaceholder extends StatelessWidget {
-  final String questionText;
+  final String? questionText;
   QuestionPlaceholder(this.questionText);
   @override
   Widget build(BuildContext context) {
@@ -10,7 +10,7 @@ class QuestionPlaceholder extends StatelessWidget {
       margin: const EdgeInsets.all(8.0),
       padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
       child: Text(questionText ?? '',
-        style: Theme.of(context).textTheme.headline6.copyWith(
+        style: Theme.of(context).textTheme.headline6!.copyWith(
           fontSize: 20,
         ),
       ),
