@@ -19,7 +19,7 @@ class _TimerState extends State<Timer> {
         decoration: BoxDecoration(
           boxShadow: [
             BoxShadow(
-              color: theme.darkTheme ? Colors.black : Colors.indigo[200],
+              color: theme.darkTheme ? Colors.black : Colors.indigo[200]!,
               spreadRadius: 3,
               blurRadius: 8,
               offset: Offset(-2, 3),
